course_2/module4_efficient_training_pipeline/assignment/requirements.txt
dlai_grader
ipywidgets==8.1.2
lightning==2.5.3
numpy==2.3.2
torch==2.7.0
torchmetrics==1.7.2
torchvision==0.22.0
matplotlib==3.10.3
seaborn==0.13.2
pillow==10.2.0
tqdm==4.67.1